modernize TeamMembershipForm
* typescript
* QueryLoader
* No more redux
